+++ src/lib.rs
@@ -81,15 +81,15 @@
 /// ```
 pub mod block {
     pub use lz4rip_core::{CompressError, DecompressError};
-    pub use lz4rip_decode::{decompress_into, decompress_into_with_dict, Decompressor};
+    pub use lz4rip_decode::{decompress_into, decompress_into_with_dict, DecompressorRef};
     pub use lz4rip_encode::{
-        compress_into, compress_into_with_dict, get_maximum_output_size, Compressor,
+        compress_into, compress_into_with_dict, get_maximum_output_size, CompressorRef,
     };
 
     #[cfg(feature = "alloc")]
-    pub use lz4rip_decode::decompress;
+    pub use lz4rip_decode::{decompress, Decompressor};
     #[cfg(feature = "alloc")]
-    pub use lz4rip_encode::{compress, DictTrainer};
+    pub use lz4rip_encode::{compress, Compressor, DictTrainer};
 }
 
 #[cfg(feature = "frame")]